RiveScript session based instance provider implemented
